+++ CHANGELOG.md
@@ -46,6 +46,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [CSE] CSEBase.srt attribute now also returns the announced attributes.
 - [CSE] Improved shutdown behaviour. Waiting for internal threads to finish.
 - [CSE] When updating a subscription resource's *nu* attribute: Removed URI's don't cause a "deletion notification" anymore.
+- [CSE] Improved *creator* attribute handling during CREATE.
 - [ACP] Added check that *pvs* is not empty during CREATE or UPDATE.
 - [ACP] Added check that 'acpi', if present, is the only attribute in an UPDATE request.